backend/requirements.txt
fastapi
uvicorn[standard]
boto3
GitPython
python-dotenv
